Changed sendCommand to send bytes

+++ lib/ME200S/ME200S.hpp
@@ -20,7 +20,7 @@ private:
 
     char commandReplyStorage[MAX_REPLAY_SIZE];
     Vector<char> receivedData_vec;
-    bool commandReplay(int commandType, int Paramater_Size);
+    bool commandReplay(unsigned long  commandType, int Paramater_Size);
